@@ -2,34 +2,46 @@
   "name": "NavigationPlayground",
   "version": "0.1.0",
   "private": true,
-  "devDependencies": {
-    "babel-jest": "^20.0.1",
-    "jest": "^20.0.1",
-    "jest-expo": "^0.4.0",
-    "react-navigation": "file:../..",
-    "react-test-renderer": "15.5.4"
-  },
-  "main": "./examples/node_modules/react-native-scripts/build/bin/crna-entry.js",
+  "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
   "scripts": {
     "start": "react-native-scripts start",
     "eject": "react-native-scripts eject",
     "android": "react-native-scripts android",
     "ios": "react-native-scripts ios",
-    "test": "node node_modules/jest/bin/jest.js"
+    "test": "flow"
   },
   "dependencies": {
-    "expo": "^16.0.0",
-    "react": "16.0.0-alpha.6",
-    "react-native": "^0.43.4"
+    "expo": "^26.0.0",
+    "react": "16.3.0-alpha.1",
+    "react-native": "^0.54.0",
+    "react-native-iphone-x-helper": "^1.0.2",
+    "react-navigation": "link:../..",
+    "react-navigation-header-buttons": "^0.0.4"
+  },
+  "devDependencies": {
+    "babel-jest": "^22.4.1",
+    "babel-plugin-transform-remove-console": "^6.9.0",
+    "flow-bin": "^0.61.0",
+    "jest": "^22.1.3",
+    "jest-expo": "^26.0.0",
+    "react-native-scripts": "^1.5.0",
+    "react-test-renderer": "16.3.0-alpha.1"
   },
   "jest": {
     "preset": "jest-expo",
+    "moduleFileExtensions": [
+      "web.js",
+      "js",
+      "json",
+      "jsx",
+      "node"
+    ],
     "modulePathIgnorePatterns": [
-      "/node_modules\/.*\/react-native/",
-      "/node_modules\/.*\/react/"
+      "/node_modules/.*/react-native/",
+      "/node_modules/.*/react/"
     ],
     "transformIgnorePatterns": [
-      "/node_modules\/(?!react-native|react-navigation)/"
+      "/node_modules/(?!react-native|react-navigation)/"
     ]
   }
 }
