@@ -2,33 +2,39 @@
   "name": "ReduxExample",
   "version": "0.0.1",
   "private": true,
+  "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
   "scripts": {
-    "test": "jest"
-  },
-  "dependencies": {
-    "prop-types": "^15.5.10",
-    "react": "16.0.0-alpha.6",
-    "react-native": "^0.44.0",
-    "react-redux": "5.0.4",
-    "redux": "3.6.0"
-  },
-  "devDependencies": {
-    "babel-jest": "^20.0.1",
-    "jest": "^20.0.1",
-    "react-navigation": "file:../..",
-    "react-test-renderer": "15.5.4"
+    "start": "react-native-scripts start",
+    "eject": "react-native-scripts eject",
+    "android": "react-native-scripts android",
+    "ios": "react-native-scripts ios",
+    "test": "node node_modules/jest/bin/jest.js"
   },
   "jest": {
-    "preset": "react-native",
-    "setupFiles": [
-      "<rootDir>/../../jest-setup.js"
-    ],
+    "preset": "jest-expo",
     "modulePathIgnorePatterns": [
-      "/node_modules\/.*\/react-native/",
-      "/node_modules\/.*\/react/"
+      "/node_modules/.*/react-native/",
+      "/node_modules/.*/react/"
     ],
     "transformIgnorePatterns": [
-      "/node_modules\/(?!react-native|react-navigation)/"
+      "/node_modules/(?!react-native|react-navigation)/"
     ]
+  },
+  "dependencies": {
+    "expo": "^25.0.0",
+    "prop-types": "^15.5.10",
+    "react": "16.2.0",
+    "react-native": "^0.52.0",
+    "react-navigation": "link:../..",
+    "react-navigation-redux-helpers": "^1.0.0",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2"
+  },
+  "devDependencies": {
+    "babel-jest": "^22.4.1",
+    "jest": "^22.1.3",
+    "jest-expo": "^25.1.0",
+    "react-native-scripts": "^1.3.1",
+    "react-test-renderer": "16.0.0"
   }
 }
