@@ -1,99 +1,102 @@
 {
   "name": "react-navigation",
-  "version": "1.0.0-beta.10",
-  "description": "React Navigation",
+  "version": "2.0.0-rc.7",
+  "description": "Routing and navigation for your React Native apps",
   "main": "src/react-navigation.js",
-  "sources": {
-    "react-native-v1": "lib-rn/react-navigation.js",
-    "web": "lib/react-navigation.web.js"
-  },
-  "module": "lib/react-navigation.web.js",
-  "webpack": "lib/react-navigation.web.js",
   "repository": {
-    "url": "git@github.com:react-community/react-navigation.git",
+    "url": "git@github.com:react-navigation/react-navigation.git",
     "type": "git"
   },
-  "author": "Adam Miskiewicz <adam@sk3vy.com>",
+  "author": "Adam Miskiewicz <adam@sk3vy.com>, Eric Vicenti <ericvicenti@gmail.com>, Brent Vatne <brent@expo.io>",
   "license": "BSD-2-Clause",
   "scripts": {
-    "clean": "rm -rf lib/ lib-rn/",
-    "build": "npm run build-docs && npm run build-web && npm run build-rn && npm run build-dot-flow",
-    "watch-rn": "mkdir -p lib-rn/views/ && cp -R src/views/assets lib/views/assets && BABEL_ENV=publish-rn babel -w src -d lib-rn",
-    "watch-web": "mkdir -p lib/views/ &&  cp -R src/views/assets lib/views/assets && BABEL_ENV=publish-web babel -w src -d lib",
-    "build-rn": "mkdir -p lib-rn/views/ && cp -R src/views/assets lib-rn/views/assets && BABEL_ENV=publish-rn babel src -d lib-rn --ignore __tests__",
-    "build-web": "mkdir -p lib/views/ && cp -R src/views/assets lib/views/assets && BABEL_ENV=publish-web babel src -d lib --ignore __tests__",
-    "build-dot-flow": "find ./src -name '*.js' -not -path '*/__tests__*' | while read filepath; do cp $filepath `echo $filepath | sed 's/\\/src\\//\\/lib\\//g'`.flow && cp $filepath `echo $filepath | sed 's/\\/src\\//\\/lib-rn\\//g'`.flow; done",
-    "build-docs": "node scripts/compile-docs",
-    "start": "npm run run-playground-ios",
-    "run-playground-ios": "cd examples/NavigationPlayground && npm run ios",
-    "run-playground-android": "cd examples/NavigationPlayground && npm run android",
-    "test": "npm run lint && npm run flow-check && npm run jest",
+    "start": "npm run ios",
+    "ios": "cd examples/NavigationPlayground && yarn && yarn ios",
+    "android": "cd examples/NavigationPlayground && yarn && yarn android",
+    "test": "npm run lint && npm run jest",
     "codecov": "codecov",
     "jest": "jest",
+    "test-update-snapshot": "jest --updateSnapshot",
     "lint": "eslint .",
     "format": "eslint --fix .",
-    "flow": "flow",
-    "flow-check": "flow check --show-all-errors",
-    "prepublish": "npm run clean && npm run build"
+    "precommit": "lint-staged"
   },
   "files": [
-    "lib",
-    "lib-rn",
     "src"
   ],
-  "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^20.0.1",
-    "babel-plugin-flow-react-proptypes": "^2.2.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-native": "^1.9.2",
-    "babel-preset-react-native-stage-0": "^1.0.1",
-    "babel-preset-react-native-syntax": "^1.0.0",
-    "babel-preset-stage-1": "^6.24.1",
-    "codecov": "^2.2.0",
-    "eslint": "^3.19.0",
-    "eslint-config-prettier": "^2.1.0",
-    "eslint-plugin-flowtype": "^2.33.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^5.0.1",
-    "eslint-plugin-prettier": "^2.0.1",
-    "eslint-plugin-react": "^7.0.1",
-    "flow-bin": "0.42.0",
-    "jest": "^20.0.1",
-    "prettier": "^1.3.1",
-    "prettier-eslint": "^6.2.2",
-    "react": "16.0.0-alpha.6",
-    "react-native": "^0.44.0",
-    "react-native-vector-icons": "^4.1.1",
-    "react-test-renderer": "^15.4.2"
-  },
   "peerDependencies": {
     "react": "*",
     "react-native": "*"
   },
   "dependencies": {
     "clamp": "^1.0.1",
-    "fbjs": "^0.8.12",
-    "hoist-non-react-statics": "^1.2.0",
+    "create-react-context": "^0.2.1",
+    "hoist-non-react-statics": "^2.2.0",
     "path-to-regexp": "^1.7.0",
     "prop-types": "^15.5.10",
-    "react-native-drawer-layout-polyfill": "^1.3.0",
-    "react-native-tab-view": "^0.0.65"
+    "react-lifecycles-compat": "^1.0.2",
+    "react-native-drawer-layout-polyfill": "^1.3.2",
+    "react-native-safe-area-view": "^0.7.0",
+    "react-navigation-deprecated-tab-navigator": "1.1.0",
+    "react-navigation-tabs": "0.1.0-alpha.8"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.25.0",
+    "babel-eslint": "^7.2.3",
+    "babel-jest": "^22.4.1",
+    "babel-preset-react-native": "^2.1.0",
+    "codecov": "^2.2.0",
+    "eslint": "^4.2.0",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-react": "^7.1.0",
+    "husky": "^0.14.3",
+    "jest": "^22.1.3",
+    "jest-expo": "^25.1.0",
+    "lint-staged": "^4.2.1",
+    "prettier": "^1.12.1",
+    "prettier-eslint": "^8.8.1",
+    "react": "16.2.0",
+    "react-native": "^0.52.0",
+    "react-native-vector-icons": "^4.2.0",
+    "react-test-renderer": "^16.0.0"
   },
   "jest": {
     "notify": true,
     "preset": "react-native",
-    "testRegex": "./src/.*\\-test\\.js$",
+    "testRegex": "/__tests__/[^/]+-test\\.js$",
     "setupFiles": [
       "<rootDir>/jest-setup.js"
     ],
     "coverageDirectory": "./coverage/",
     "collectCoverage": true,
+    "coverageReporters": [
+      "lcov"
+    ],
+    "collectCoverageFrom": [
+      "src/**/*.js"
+    ],
+    "coveragePathIgnorePatterns": [
+      "jest-setup.js"
+    ],
+    "moduleNameMapper": {
+      "\\.png$": "<rootDir>/assetsTransformer.js"
+    },
     "modulePathIgnorePatterns": [
-      "examples"
+      "<rootDir>/examples/"
+    ],
+    "transformIgnorePatterns": [
+      "node_modules/(?!(jest-)?react-native|react-clone-referenced-element|react-navigation-deprecated-tab-navigator)"
+    ]
+  },
+  "lint-staged": {
+    "*.js": [
+      "eslint --fix",
+      "prettier --write flow/react-navigation.js",
+      "git add"
     ]
   }
 }
