@@ -1,33 +1,53 @@
-/* @flow */
-
 import React from 'react';
-import invariant from 'fbjs/lib/invariant';
-import { BackAndroid, Linking } from './PlatformHelpers';
+import { Linking, AsyncStorage } from 'react-native';
+import withLifecyclePolyfill from 'react-lifecycles-compat';
+
+import { BackHandler } from './PlatformHelpers';
 import NavigationActions from './NavigationActions';
-import addNavigationHelpers from './addNavigationHelpers';
+import invariant from './utils/invariant';
+import docsUrl from './utils/docsUrl';
 
-import type {
-  NavigationRoute,
-  NavigationAction,
-  NavigationState,
-  NavigationScreenProp,
-  NavigationNavigatorProps,
-} from './TypeDefinition';
+function isStateful(props) {
+  return !props.navigation;
+}
 
-type NavigationContainerProps = {
-  uriPrefix?: string | RegExp,
-  onNavigationStateChange?: (
-    NavigationState,
-    NavigationState,
-    NavigationAction
-  ) => void,
-};
+function validateProps(props) {
+  if (isStateful(props)) {
+    return;
+  }
 
-type Props<T> = NavigationContainerProps & NavigationNavigatorProps<T>;
+  const { navigation, screenProps, ...containerProps } = props;
 
-type State = {
-  nav: ?NavigationState,
-};
+  const keys = Object.keys(containerProps);
+
+  if (keys.length !== 0) {
+    throw new Error(
+      'This navigator has both navigation and container props, so it is ' +
+        `unclear if it should own its own state. Remove props: "${keys.join(
+          ', '
+        )}" ` +
+        'if the navigator should get its state from the navigation prop. If the ' +
+        'navigator should maintain its own state, do not pass a navigation prop.'
+    );
+  }
+}
+
+// Track the number of stateful container instances. Warn if >0 and not using the
+// detached prop to explicitly acknowledge the behavior. We should deprecated implicit
+// stateful navigation containers in a future release and require a provider style pattern
+// instead in order to eliminate confusion entirely.
+let _statefulContainerCount = 0;
+export function _TESTING_ONLY_reset_container_count() {
+  _statefulContainerCount = 0;
+}
+
+// We keep a global flag to catch errors during the state persistence hydrating scenario.
+// The innermost navigator who catches the error will dispatch a new init action.
+let _reactNavigationIsHydratingState = false;
+// Unfortunate to use global state here, but it seems necessesary for the time
+// being. There seems to be some problems with cascading componentDidCatch
+// handlers. Ideally the inner non-stateful navigator catches the error and
+// re-throws it, to be caught by the top-level stateful navigator.
 
 /**
  * Create an HOC that injects the navigation and manages the navigation state
@@ -35,42 +55,59 @@ type State = {
  * This allows to use e.g. the StackNavigator and TabNavigator as root-level
  * components.
  */
-export default function createNavigationContainer<T: *>(
-  Component: ReactClass<NavigationNavigatorProps<T>>,
-  containerOptions?: {}
-) {
-  invariant(
-    typeof containerOptions === 'undefined',
-    'containerOptions.URIPrefix has been removed. Pass the uriPrefix prop to the navigator instead'
-  );
-
-  class NavigationContainer extends React.Component<void, Props<T>, State> {
-    state: State;
-    props: Props<T>;
-
-    subs: ?{
-      remove: () => void,
-    } = null;
+export default function createNavigationContainer(Component) {
+  class NavigationContainer extends React.Component {
+    subs = null;
 
     static router = Component.router;
+    static navigationOptions = null;
 
-    constructor(props: Props<T>) {
+    static getDerivedStateFromProps(nextProps, prevState) {
+      validateProps(nextProps);
+      return null;
+    }
+
+    _actionEventSubscribers = new Set();
+
+    constructor(props) {
       super(props);
 
-      this._validateProps(props);
+      validateProps(props);
+
+      this._initialAction = NavigationActions.init();
+
+      if (this._isStateful()) {
+        this.subs = BackHandler.addEventListener('hardwareBackPress', () => {
+          if (!this._isMounted) {
+            this.subs && this.subs.remove();
+          } else {
+            // dispatch returns true if the action results in a state change,
+            // and false otherwise. This maps well to what BackHandler expects
+            // from a callback -- true if handled, false if not handled
+            return this.dispatch(NavigationActions.back());
+          }
+        });
+      }
 
       this.state = {
-        nav: this._isStateful()
-          ? Component.router.getStateForAction(NavigationActions.init())
-          : null,
+        nav:
+          this._isStateful() && !props.persistenceKey
+            ? Component.router.getStateForAction(this._initialAction)
+            : null,
       };
     }
 
-    _isStateful(): boolean {
-      return !this.props.navigation;
+    _renderLoading() {
+      return this.props.renderLoadingExperimental
+        ? this.props.renderLoadingExperimental()
+        : null;
     }
 
-    _validateProps(props: Props<T>) {
+    _isStateful() {
+      return isStateful(this.props);
+    }
+
+    _validateProps(props) {
       if (this._isStateful()) {
         return;
       }
@@ -79,21 +116,26 @@ export default function createNavigationContainer<T: *>(
 
       const keys = Object.keys(containerProps);
 
-      invariant(
-        keys.length === 0,
-        'This navigator has both navigation and container props, so it is ' +
-          `unclear if it should own its own state. Remove props: "${keys.join(', ')}" ` +
-          'if the navigator should get its state from the navigation prop. If the ' +
-          'navigator should maintain its own state, do not pass a navigation prop.'
-      );
+      if (keys.length !== 0) {
+        throw new Error(
+          'This navigator has both navigation and container props, so it is ' +
+            `unclear if it should own its own state. Remove props: "${keys.join(
+              ', '
+            )}" ` +
+            'if the navigator should get its state from the navigation prop. If the ' +
+            'navigator should maintain its own state, do not pass a navigation prop.'
+        );
+      }
     }
 
-    _urlToPathAndParams(url: string) {
+    _urlToPathAndParams(url) {
       const params = {};
       const delimiter = this.props.uriPrefix || '://';
       let path = url.split(delimiter)[1];
-      if (!path) {
+      if (typeof path === 'undefined') {
         path = url;
+      } else if (path === '') {
+        path = '/';
       }
       return {
         path,
@@ -101,7 +143,7 @@ export default function createNavigationContainer<T: *>(
       };
     }
 
-    _handleOpenURL = (url: string) => {
+    _handleOpenURL = ({ url }) => {
       const parsedUrl = this._urlToPathAndParams(url);
       if (parsedUrl) {
         const { path, params } = parsedUrl;
@@ -112,14 +154,11 @@ export default function createNavigationContainer<T: *>(
       }
     };
 
-    _onNavigationStateChange(
-      prevNav: NavigationState,
-      nav: NavigationState,
-      action: NavigationAction
-    ) {
+    _onNavigationStateChange(prevNav, nav, action) {
       if (
         typeof this.props.onNavigationStateChange === 'undefined' &&
-        this._isStateful()
+        this._isStateful() &&
+        !!process.env.REACT_NAV_LOGGING
       ) {
         /* eslint-disable no-console */
         if (console.group) {
@@ -144,64 +183,189 @@ export default function createNavigationContainer<T: *>(
       }
     }
 
-    componentWillReceiveProps(nextProps: *) {
-      this._validateProps(nextProps);
+    componentDidUpdate() {
+      // Clear cached _nav every tick
+      if (this._nav === this.state.nav) {
+        this._nav = null;
+      }
     }
 
-    componentDidMount() {
+    async componentDidMount() {
+      this._isMounted = true;
       if (!this._isStateful()) {
         return;
       }
 
-      this.subs = BackAndroid.addEventListener('backPress', () =>
-        this.dispatch(NavigationActions.back())
-      );
+      if (__DEV__ && !this.props.detached) {
+        if (_statefulContainerCount > 0) {
+          console.error(
+            `You should only render one navigator explicitly in your app, and other navigators should by rendered by including them in that navigator. Full details at: ${docsUrl(
+              'common-mistakes.html#explicitly-rendering-more-than-one-navigator'
+            )}`
+          );
+        }
+      }
+      _statefulContainerCount++;
+      Linking.addEventListener('url', this._handleOpenURL);
 
-      Linking.addEventListener('url', ({ url }: { url: string }) => {
-        this._handleOpenURL(url);
+      // Pull out anything that can impact state
+      const { persistenceKey } = this.props;
+      const startupStateJSON =
+        persistenceKey && (await AsyncStorage.getItem(persistenceKey));
+      const url = await Linking.getInitialURL();
+      const parsedUrl = url && this._urlToPathAndParams(url);
+
+      // Initialize state. This must be done *after* any async code
+      // so we don't end up with a different value for this.state.nav
+      // due to changes while async function was resolving
+      let action = this._initialAction;
+      let startupState = this.state.nav;
+      if (!startupState) {
+        !!process.env.REACT_NAV_LOGGING &&
+          console.log('Init new Navigation State');
+        startupState = Component.router.getStateForAction(action);
+      }
+
+      // Pull persisted state from AsyncStorage
+      if (startupStateJSON) {
+        try {
+          startupState = JSON.parse(startupStateJSON);
+          _reactNavigationIsHydratingState = true;
+        } catch (e) {}
+      }
+
+      // Pull state out of URL
+      if (parsedUrl) {
+        const { path, params } = parsedUrl;
+        const urlAction = Component.router.getActionForPathAndParams(
+          path,
+          params
+        );
+        if (urlAction) {
+          !!process.env.REACT_NAV_LOGGING &&
+            console.log('Applying Navigation Action for Initial URL:', url);
+          action = urlAction;
+          startupState = Component.router.getStateForAction(
+            urlAction,
+            startupState
+          );
+        }
+      }
+
+      const dispatchActions = () =>
+        this._actionEventSubscribers.forEach(subscriber =>
+          subscriber({
+            type: 'action',
+            action,
+            state: this.state.nav,
+            lastState: null,
+          })
+        );
+
+      if (startupState === this.state.nav) {
+        dispatchActions();
+        return;
+      }
+
+      this.setState({ nav: startupState }, () => {
+        _reactNavigationIsHydratingState = false;
+        dispatchActions();
       });
-
-      Linking.getInitialURL().then(
-        (url: string) => url && this._handleOpenURL(url)
-      );
     }
+
+    componentDidCatch(e, errorInfo) {
+      if (_reactNavigationIsHydratingState) {
+        _reactNavigationIsHydratingState = false;
+        console.warn(
+          'Uncaught exception while starting app from persisted navigation state! Trying to render again with a fresh navigation state..'
+        );
+        this.dispatch(NavigationActions.init());
+      }
+    }
+
+    _persistNavigationState = async nav => {
+      const { persistenceKey } = this.props;
+      if (!persistenceKey) {
+        return;
+      }
+      await AsyncStorage.setItem(persistenceKey, JSON.stringify(nav));
+    };
 
     componentWillUnmount() {
+      this._isMounted = false;
       Linking.removeEventListener('url', this._handleOpenURL);
       this.subs && this.subs.remove();
+
+      if (this._isStateful()) {
+        _statefulContainerCount--;
+      }
     }
 
-    dispatch = (action: NavigationAction) => {
-      const { state } = this;
-      if (!this._isStateful()) {
-        return false;
+    // Per-tick temporary storage for state.nav
+
+    dispatch = action => {
+      if (this.props.navigation) {
+        return this.props.navigation.dispatch(action);
       }
-      const nav = Component.router.getStateForAction(action, state.nav);
-      if (nav && nav !== state.nav) {
-        this.setState({ nav }, () =>
-          this._onNavigationStateChange(state.nav, nav, action)
+      this._nav = this._nav || this.state.nav;
+      const oldNav = this._nav;
+      invariant(oldNav, 'should be set in constructor if stateful');
+      const nav = Component.router.getStateForAction(action, oldNav);
+      const dispatchActionEvents = () => {
+        this._actionEventSubscribers.forEach(subscriber =>
+          subscriber({
+            type: 'action',
+            action,
+            state: nav,
+            lastState: oldNav,
+          })
         );
+      };
+      if (nav && nav !== oldNav) {
+        // Cache updates to state.nav during the tick to ensure that subsequent calls will not discard this change
+        this._nav = nav;
+        this.setState({ nav }, () => {
+          this._onNavigationStateChange(oldNav, nav, action);
+          dispatchActionEvents();
+          this._persistNavigationState(nav);
+        });
         return true;
+      } else {
+        dispatchActionEvents();
       }
       return false;
     };
 
-    _navigation: ?NavigationScreenProp<NavigationRoute, NavigationAction>;
-
     render() {
       let navigation = this.props.navigation;
       if (this._isStateful()) {
-        if (!this._navigation || this._navigation.state !== this.state.nav) {
-          this._navigation = addNavigationHelpers({
+        const nav = this.state.nav;
+        if (!nav) {
+          return this._renderLoading();
+        }
+        if (!this._navigation || this._navigation.state !== nav) {
+          this._navigation = {
             dispatch: this.dispatch,
-            state: this.state.nav,
-          });
+            state: nav,
+            addListener: (eventName, handler) => {
+              if (eventName !== 'action') {
+                return { remove: () => {} };
+              }
+              this._actionEventSubscribers.add(handler);
+              return {
+                remove: () => {
+                  this._actionEventSubscribers.delete(handler);
+                },
+              };
+            },
+          };
         }
         navigation = this._navigation;
       }
+      invariant(navigation, 'failed to get navigation');
       return <Component {...this.props} navigation={navigation} />;
     }
   }
 
-  return NavigationContainer;
+  return withLifecyclePolyfill(NavigationContainer);
 }
