@@ -1,155 +1,118 @@
-/* @flow */
-
-import React, { PureComponent } from 'react';
+import React from 'react';
+import { Dimensions } from 'react-native';
 import DrawerLayout from 'react-native-drawer-layout-polyfill';
 
-import addNavigationHelpers from '../../addNavigationHelpers';
 import DrawerSidebar from './DrawerSidebar';
-
-import type {
-  NavigationScreenProp,
-  NavigationRoute,
-  NavigationRouter,
-  NavigationState,
-  NavigationAction,
-  NavigationDrawerScreenOptions,
-  Style,
-} from '../../TypeDefinition';
-
-export type DrawerScene = {
-  route: NavigationRoute,
-  focused: boolean,
-  index: number,
-  tintColor?: string,
-};
-
-export type DrawerItem = {
-  route: NavigationRoute,
-  focused: boolean,
-};
-
-export type DrawerViewConfig = {
-  drawerWidth: number,
-  drawerPosition: 'left' | 'right',
-  contentComponent: ReactClass<*>,
-  contentOptions?: {},
-  style?: Style,
-};
-
-type Props = DrawerViewConfig & {
-  screenProps?: {},
-  router: NavigationRouter<
-    NavigationState,
-    NavigationAction,
-    NavigationDrawerScreenOptions
-  >,
-  navigation: NavigationScreenProp<NavigationState, NavigationAction>,
-};
+import NavigationActions from '../../NavigationActions';
+import DrawerActions from '../../routers/DrawerActions';
 
 /**
  * Component that renders the drawer.
  */
-export default class DrawerView<T: *> extends PureComponent<void, Props, void> {
-  props: Props;
+export default class DrawerView extends React.PureComponent {
+  state = {
+    drawerWidth:
+      typeof this.props.navigationConfig.drawerWidth === 'function'
+        ? this.props.navigationConfig.drawerWidth()
+        : this.props.navigationConfig.drawerWidth,
+  };
 
-  componentWillMount() {
-    this._updateScreenNavigation(this.props.navigation);
+  componentDidMount() {
+    Dimensions.addEventListener('change', this._updateWidth);
   }
 
-  componentWillReceiveProps(nextProps: Props) {
-    if (
-      this.props.navigation.state.index !== nextProps.navigation.state.index
-    ) {
-      const { routes, index } = nextProps.navigation.state;
-      if (routes[index].routeName === 'DrawerOpen') {
-        this._drawer.openDrawer();
-      } else {
-        this._drawer.closeDrawer();
-      }
+  componentWillUnmount() {
+    Dimensions.removeEventListener('change', this._updateWidth);
+  }
+
+  componentDidUpdate(prevProps, prevState) {
+    const { isDrawerOpen } = this.props.navigation.state;
+    const wasDrawerOpen = prevProps.navigation.state.isDrawerOpen;
+
+    if (isDrawerOpen && !wasDrawerOpen) {
+      this._drawer.openDrawer();
+    } else if (wasDrawerOpen && !isDrawerOpen) {
+      this._drawer.closeDrawer();
     }
-    this._updateScreenNavigation(nextProps.navigation);
   }
-
-  _screenNavigationProp: NavigationScreenProp<T, NavigationAction>;
 
   _handleDrawerOpen = () => {
     const { navigation } = this.props;
-    const { routes, index } = navigation.state;
-    if (routes[index].routeName !== 'DrawerOpen') {
-      this.props.navigation.navigate('DrawerOpen');
+    const { isDrawerOpen } = navigation.state;
+    if (!isDrawerOpen) {
+      navigation.dispatch({ type: DrawerActions.OPEN_DRAWER });
     }
   };
 
   _handleDrawerClose = () => {
     const { navigation } = this.props;
-    const { routes, index } = navigation.state;
-    if (routes[index].routeName !== 'DrawerClose') {
-      this.props.navigation.navigate('DrawerClose');
+    const { isDrawerOpen } = navigation.state;
+    if (isDrawerOpen) {
+      navigation.dispatch({ type: DrawerActions.CLOSE_DRAWER });
     }
   };
 
-  _updateScreenNavigation = (
-    navigation: NavigationScreenProp<NavigationState, NavigationAction>
-  ) => {
-    const navigationState = navigation.state.routes.find(
-      (route: *) => route.routeName === 'DrawerClose'
-    );
-    if (
-      this._screenNavigationProp &&
-      this._screenNavigationProp.state === navigationState
-    ) {
-      return;
+  _updateWidth = () => {
+    const drawerWidth =
+      typeof this.props.navigationConfig.drawerWidth === 'function'
+        ? this.props.navigationConfig.drawerWidth()
+        : this.props.navigationConfig.drawerWidth;
+
+    if (this.state.drawerWidth !== drawerWidth) {
+      this.setState({ drawerWidth });
     }
-    this._screenNavigationProp = addNavigationHelpers({
-      ...navigation,
-      state: navigationState,
-    });
   };
 
-  _getNavigationState = (
-    navigation: NavigationScreenProp<NavigationState, NavigationAction>
-  ) => {
-    const navigationState = navigation.state.routes.find(
-      (route: *) => route.routeName === 'DrawerClose'
+  _renderNavigationView = () => {
+    return (
+      <DrawerSidebar
+        screenProps={this.props.screenProps}
+        navigation={this.props.navigation}
+        descriptors={this.props.descriptors}
+        contentComponent={this.props.navigationConfig.contentComponent}
+        contentOptions={this.props.navigationConfig.contentOptions}
+        drawerPosition={this.props.navigationConfig.drawerPosition}
+        style={this.props.navigationConfig.style}
+        {...this.props.navigationConfig}
+      />
     );
-    return navigationState;
   };
-
-  _renderNavigationView = () => (
-    <DrawerSidebar
-      screenProps={this.props.screenProps}
-      navigation={this._screenNavigationProp}
-      router={this.props.router}
-      contentComponent={this.props.contentComponent}
-      contentOptions={this.props.contentOptions}
-      style={this.props.style}
-    />
-  );
-
-  _drawer: any;
 
   render() {
-    const DrawerScreen = this.props.router.getComponentForRouteName(
-      'DrawerClose'
-    );
+    const { state } = this.props.navigation;
+    const activeKey = state.routes[state.index].key;
+    const descriptor = this.props.descriptors[activeKey];
+
+    const DrawerScreen = descriptor.getComponent();
+
+    const { drawerLockMode } = descriptor.options;
+
     return (
       <DrawerLayout
-        ref={(c: *) => {
+        ref={c => {
           this._drawer = c;
         }}
-        drawerWidth={this.props.drawerWidth}
+        drawerLockMode={
+          (this.props.screenProps && this.props.screenProps.drawerLockMode) ||
+          this.props.navigationConfig.drawerLockMode
+        }
+        drawerBackgroundColor={
+          this.props.navigationConfig.drawerBackgroundColor
+        }
+        drawerWidth={this.state.drawerWidth}
         onDrawerOpen={this._handleDrawerOpen}
         onDrawerClose={this._handleDrawerClose}
+        useNativeAnimations={this.props.navigationConfig.useNativeAnimations}
         renderNavigationView={this._renderNavigationView}
         drawerPosition={
-          this.props.drawerPosition === 'right'
+          this.props.navigationConfig.drawerPosition === 'right'
             ? DrawerLayout.positions.Right
             : DrawerLayout.positions.Left
         }
       >
         <DrawerScreen
           screenProps={this.props.screenProps}
-          navigation={this._screenNavigationProp}
+          navigation={descriptor.navigation}
         />
       </DrawerLayout>
     );
