@@ -1,65 +1,18 @@
-/* @flow */
-
 import React from 'react';
-
-import { Animated, StyleSheet, View } from 'react-native';
-
-import invariant from 'fbjs/lib/invariant';
+import { Animated, Easing, StyleSheet, View } from 'react-native';
+import invariant from '../utils/invariant';
 
 import NavigationScenesReducer from './ScenesReducer';
-import TransitionConfigs from './TransitionConfigs';
 
-import type {
-  NavigationLayout,
-  NavigationScene,
-  NavigationState,
-  NavigationAction,
-  NavigationScreenProp,
-  NavigationTransitionProps,
-  NavigationTransitionSpec,
-} from '../TypeDefinition';
-
-type Props = {
-  configureTransition: (
-    transitionProps: NavigationTransitionProps,
-    prevTransitionProps: ?NavigationTransitionProps
-  ) => NavigationTransitionSpec,
-  navigation: NavigationScreenProp<NavigationState, NavigationAction>,
-  onTransitionEnd?: () => void,
-  onTransitionStart?: () => void,
-  render: (
-    transitionProps: NavigationTransitionProps,
-    prevTransitionProps: ?NavigationTransitionProps
-  ) => any,
-  style?: any,
+// Used for all animations unless overriden
+const DefaultTransitionSpec = {
+  duration: 250,
+  easing: Easing.inOut(Easing.ease),
+  timing: Animated.timing,
 };
 
-type State = {
-  layout: NavigationLayout,
-  position: Animated.Value,
-  progress: Animated.Value,
-  scenes: Array<NavigationScene>,
-};
-
-const DefaultTransitionSpec = TransitionConfigs.DefaultTransitionSpec;
-
-class Transitioner extends React.Component<*, Props, State> {
-  _onLayout: (event: any) => void;
-  _onTransitionEnd: () => void;
-  _prevTransitionProps: ?NavigationTransitionProps;
-  _transitionProps: NavigationTransitionProps;
-  _isMounted: boolean;
-  _isTransitionRunning: boolean;
-  _queuedTransition: ?{
-    nextProps: Props,
-    nextScenes: Array<NavigationScene>,
-    indexHasChanged: boolean,
-  };
-
-  props: Props;
-  state: State;
-
-  constructor(props: Props, context: any) {
+class Transitioner extends React.Component {
+  constructor(props, context) {
     super(props, context);
 
     // The initial layout isn't measured. Measured layout will be only available
@@ -76,7 +29,12 @@ class Transitioner extends React.Component<*, Props, State> {
       layout,
       position: new Animated.Value(this.props.navigation.state.index),
       progress: new Animated.Value(1),
-      scenes: NavigationScenesReducer([], this.props.navigation.state),
+      scenes: NavigationScenesReducer(
+        [],
+        this.props.navigation.state,
+        null,
+        this.props.descriptors
+      ),
     };
 
     this._prevTransitionProps = null;
@@ -86,25 +44,24 @@ class Transitioner extends React.Component<*, Props, State> {
     this._queuedTransition = null;
   }
 
-  componentWillMount(): void {
-    this._onLayout = this._onLayout.bind(this);
-    this._onTransitionEnd = this._onTransitionEnd.bind(this);
-  }
-
-  componentDidMount(): void {
+  componentDidMount() {
     this._isMounted = true;
   }
 
-  componentWillUnmount(): void {
+  componentWillUnmount() {
     this._isMounted = false;
   }
 
-  componentWillReceiveProps(nextProps: Props): void {
-    const nextScenes = NavigationScenesReducer(
+  componentWillReceiveProps(nextProps) {
+    let nextScenes = NavigationScenesReducer(
       this.state.scenes,
       nextProps.navigation.state,
-      this.props.navigation.state
+      this.props.navigation.state,
+      nextProps.descriptors
     );
+    if (!nextProps.navigation.state.isTransitioning) {
+      nextScenes = filterStale(nextScenes);
+    }
 
     if (nextScenes === this.state.scenes) {
       return;
@@ -120,11 +77,7 @@ class Transitioner extends React.Component<*, Props, State> {
     this._startTransition(nextProps, nextScenes, indexHasChanged);
   }
 
-  _startTransition(
-    nextProps: Props,
-    nextScenes: Array<NavigationScene>,
-    indexHasChanged: boolean
-  ) {
+  _startTransition(nextProps, nextScenes, indexHasChanged) {
     const nextState = {
       ...this.state,
       scenes: nextScenes,
@@ -157,45 +110,46 @@ class Transitioner extends React.Component<*, Props, State> {
     const positionHasChanged = position.__getValue() !== toValue;
 
     // if swiped back, indexHasChanged == true && positionHasChanged == false
-    const animations = indexHasChanged && positionHasChanged
-      ? [
-          timing(progress, {
-            ...transitionSpec,
-            toValue: 1,
-          }),
-          timing(position, {
-            ...transitionSpec,
-            toValue: nextProps.navigation.state.index,
-          }),
-        ]
-      : [];
+    const animations =
+      indexHasChanged && positionHasChanged
+        ? [
+            timing(progress, {
+              ...transitionSpec,
+              toValue: 1,
+            }),
+            timing(position, {
+              ...transitionSpec,
+              toValue: nextProps.navigation.state.index,
+            }),
+          ]
+        : [];
 
     // update scenes and play the transition
     this._isTransitionRunning = true;
-
-    const scenes = this.state.scenes;
-
-    this.setState(nextState, () => {
-      nextProps.onTransitionStart &&
-        nextProps.onTransitionStart(
+    this.setState(nextState, async () => {
+      if (nextProps.onTransitionStart) {
+        const result = nextProps.onTransitionStart(
           this._transitionProps,
-          this._prevTransitionProps,
-          this.state.scenes,
-          scenes,
+          this._prevTransitionProps
         );
+
+        if (result instanceof Promise) {
+          await result;
+        }
+      }
       Animated.parallel(animations).start(this._onTransitionEnd);
     });
   }
 
   render() {
     return (
-      <View onLayout={this._onLayout} style={[styles.main, this.props.style]}>
+      <View onLayout={this._onLayout} style={[styles.main]}>
         {this.props.render(this._transitionProps, this._prevTransitionProps)}
       </View>
     );
   }
 
-  _onLayout(event: any): void {
+  _onLayout = event => {
     const { height, width } = event.nativeEvent.layout;
     if (
       this.state.layout.initWidth === width &&
@@ -220,25 +174,36 @@ class Transitioner extends React.Component<*, Props, State> {
 
     this._transitionProps = buildTransitionProps(this.props, nextState);
     this.setState(nextState);
-  }
+  };
 
-  _onTransitionEnd(): void {
+  _onTransitionEnd = () => {
     if (!this._isMounted) {
       return;
     }
     const prevTransitionProps = this._prevTransitionProps;
     this._prevTransitionProps = null;
 
+    const scenes = filterStale(this.state.scenes);
+
     const nextState = {
       ...this.state,
-      scenes: this.state.scenes.filter(isSceneNotStale),
+      scenes,
     };
 
     this._transitionProps = buildTransitionProps(this.props, nextState);
 
-    this.setState(nextState, () => {
-      this.props.onTransitionEnd &&
-        this.props.onTransitionEnd(this._transitionProps, prevTransitionProps);
+    this.setState(nextState, async () => {
+      if (this.props.onTransitionEnd) {
+        const result = this.props.onTransitionEnd(
+          this._transitionProps,
+          prevTransitionProps
+        );
+
+        if (result instanceof Promise) {
+          await result;
+        }
+      }
+
       if (this._queuedTransition) {
         this._startTransition(
           this._queuedTransition.nextProps,
@@ -250,13 +215,10 @@ class Transitioner extends React.Component<*, Props, State> {
         this._isTransitionRunning = false;
       }
     });
-  }
+  };
 }
 
-function buildTransitionProps(
-  props: Props,
-  state: State
-): NavigationTransitionProps {
+function buildTransitionProps(props, state) {
   const { navigation } = props;
 
   const { layout, position, progress, scenes } = state;
@@ -276,11 +238,19 @@ function buildTransitionProps(
   };
 }
 
-function isSceneNotStale(scene: NavigationScene): boolean {
+function isSceneNotStale(scene) {
   return !scene.isStale;
 }
 
-function isSceneActive(scene: NavigationScene): boolean {
+function filterStale(scenes) {
+  const filtered = scenes.filter(isSceneNotStale);
+  if (filtered.length === scenes.length) {
+    return scenes;
+  }
+  return filtered;
+}
+
+function isSceneActive(scene) {
   return scene.isActive;
 }
 
