@@ -1,10 +1,10 @@
 import React from 'react';
-import 'react-native';
-import renderer from 'react-test-renderer';
-
 import App from './App';
+
+import renderer from 'react-test-renderer';
 
 it('renders without crashing', () => {
   const rendered = renderer.create(<App />).toJSON();
-  expect(rendered).toBeTruthy();
+  // Will be null because the playground uses state persistence which happens asyncronously
+  expect(rendered).toEqual(null);
 });
