@@ -3,23 +3,31 @@
 import React from 'react';
 
 import { Image, Platform, StyleSheet, Text, View } from 'react-native';
+import { SafeAreaView } from 'react-navigation';
 
 const Banner = () => (
-  <View style={styles.banner}>
-    <Image source={require('./assets/NavLogo.png')} style={styles.image} />
-    <Text style={styles.title}>React Navigation Examples</Text>
-  </View>
+  <SafeAreaView
+    style={styles.bannerContainer}
+    forceInset={{ top: 'always' }}
+  >
+    <View style={styles.banner}>
+      <Image source={require('./assets/NavLogo.png')} style={styles.image} />
+      <Text style={styles.title}>React Navigation Examples</Text>
+    </View>
+  </SafeAreaView>
 );
 
 export default Banner;
 
 const styles = StyleSheet.create({
-  banner: {
+  bannerContainer: {
     backgroundColor: '#673ab7',
+    paddingTop: 20,
+  },
+  banner: {
     flexDirection: 'row',
     alignItems: 'center',
     padding: 16,
-    marginTop: Platform.OS === 'ios' ? 20 : 0,
   },
   image: {
     width: 36,
