@@ -1,9 +1,37 @@
-/* @flow */
-
 import NavigationActions from '../NavigationActions';
 import addNavigationHelpers from '../addNavigationHelpers';
 
+const dummyEventSubscriber = (name: string, handler: (*) => void) => ({
+  remove: () => {},
+});
+
 describe('addNavigationHelpers', () => {
+  it('handles dismiss action', () => {
+    const mockedDispatch = jest
+      .fn(() => false)
+      .mockImplementationOnce(() => true);
+    const child = { key: 'A', routeName: 'Home' };
+    expect(
+      addNavigationHelpers({
+        state: child,
+        dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
+        dangerouslyGetParent: () => ({
+          state: {
+            key: 'P',
+            routeName: 'Parent',
+            routes: [child],
+          },
+        }),
+      }).dismiss()
+    ).toEqual(true);
+    expect(mockedDispatch).toBeCalledWith({
+      type: NavigationActions.BACK,
+      key: 'P',
+    });
+    expect(mockedDispatch.mock.calls.length).toBe(1);
+  });
+
   it('handles Back action', () => {
     const mockedDispatch = jest
       .fn(() => false)
@@ -12,6 +40,7 @@ describe('addNavigationHelpers', () => {
       addNavigationHelpers({
         state: { key: 'A', routeName: 'Home' },
         dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
       }).goBack('A')
     ).toEqual(true);
     expect(mockedDispatch).toBeCalledWith({
@@ -29,6 +58,7 @@ describe('addNavigationHelpers', () => {
       addNavigationHelpers({
         state: { routeName: 'Home' },
         dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
       }).goBack()
     ).toEqual(true);
     expect(mockedDispatch).toBeCalledWith({ type: NavigationActions.BACK });
@@ -43,6 +73,7 @@ describe('addNavigationHelpers', () => {
       addNavigationHelpers({
         state: { routeName: 'Home' },
         dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
       }).navigate('Profile', { name: 'Matt' })
     ).toEqual(true);
     expect(mockedDispatch).toBeCalledWith({
@@ -61,6 +92,7 @@ describe('addNavigationHelpers', () => {
       addNavigationHelpers({
         state: { key: 'B', routeName: 'Settings' },
         dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
       }).setParams({ notificationsEnabled: 'yes' })
     ).toEqual(true);
     expect(mockedDispatch).toBeCalledWith({
@@ -69,5 +101,44 @@ describe('addNavigationHelpers', () => {
       params: { notificationsEnabled: 'yes' },
     });
     expect(mockedDispatch.mock.calls.length).toBe(1);
+  });
+
+  it('handles GetParams action', () => {
+    const mockedDispatch = jest
+      .fn(() => false)
+      .mockImplementationOnce(() => true);
+    expect(
+      addNavigationHelpers({
+        state: { key: 'B', routeName: 'Settings', params: { name: 'Peter' } },
+        dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
+      }).getParam('name', 'Brent')
+    ).toEqual('Peter');
+  });
+
+  it('handles GetParams action with default param', () => {
+    const mockedDispatch = jest
+      .fn(() => false)
+      .mockImplementationOnce(() => true);
+    expect(
+      addNavigationHelpers({
+        state: { key: 'B', routeName: 'Settings' },
+        dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
+      }).getParam('name', 'Brent')
+    ).toEqual('Brent');
+  });
+
+  it('handles GetParams action with param value as null', () => {
+    const mockedDispatch = jest
+      .fn(() => false)
+      .mockImplementationOnce(() => true);
+    expect(
+      addNavigationHelpers({
+        state: { key: 'B', routeName: 'Settings', params: { name: null } },
+        dispatch: mockedDispatch,
+        addListener: dummyEventSubscriber,
+      }).getParam('name')
+    ).toEqual(null);
   });
 });
