@@ -1,7 +1,3 @@
-/*
- * @noflow - get/set properties not yet supported by flow. also `...require(x)` is broken #6560135
- */
-
 /* eslint global-require: 0 */
 
 module.exports = {
@@ -12,25 +8,71 @@ module.exports = {
   get StateUtils() {
     return require('./StateUtils').default;
   },
-  get addNavigationHelpers() {
-    return require('./addNavigationHelpers').default;
-  },
-  get NavigationActions() {
-    return require('./NavigationActions').default;
-  },
 
   // Navigators
   get createNavigator() {
     return require('./navigators/createNavigator').default;
   },
-  get StackNavigator() {
-    return require('./navigators/StackNavigator').default;
+  get createStackNavigator() {
+    return require('./navigators/createStackNavigator').default;
   },
-  get TabNavigator() {
-    return require('./navigators/TabNavigator').default;
+  get StackNavigator() {
+    console.warn(
+      'The StackNavigator function name is deprecated, please use createStackNavigator instead'
+    );
+    return require('./navigators/createStackNavigator').default;
+  },
+  get createSwitchNavigator() {
+    return require('./navigators/createSwitchNavigator').default;
+  },
+  get SwitchNavigator() {
+    console.warn(
+      'The SwitchNavigator function name is deprecated, please use createSwitchNavigator instead'
+    );
+    return require('./navigators/createSwitchNavigator').default;
+  },
+  get createDrawerNavigator() {
+    return require('./navigators/createDrawerNavigator').default;
   },
   get DrawerNavigator() {
-    return require('./navigators/DrawerNavigator').default;
+    console.warn(
+      'The DrawerNavigator function name is deprecated, please use createDrawerNavigator instead'
+    );
+    return require('./navigators/createDrawerNavigator').default;
+  },
+  get createTabNavigator() {
+    console.warn(
+      'createTabNavigator is deprecated. Please use the createBottomTabNavigator or createMaterialTopTabNavigator instead.'
+    );
+    return require('react-navigation-deprecated-tab-navigator')
+      .createTabNavigator;
+  },
+  get TabNavigator() {
+    console.warn(
+      'TabNavigator is deprecated. Please use the createBottomTabNavigator or createMaterialTopTabNavigator instead.'
+    );
+    return require('react-navigation-deprecated-tab-navigator')
+      .createTabNavigator;
+  },
+  get createBottomTabNavigator() {
+    return require('react-navigation-tabs').createBottomTabNavigator;
+  },
+  get createMaterialBottomTabNavigator() {
+    return require('react-navigation-tabs').createMaterialBottomTabNavigator;
+  },
+  get createMaterialTopTabNavigator() {
+    return require('react-navigation-tabs').createMaterialTopTabNavigator;
+  },
+
+  // Actions
+  get NavigationActions() {
+    return require('./NavigationActions').default;
+  },
+  get StackActions() {
+    return require('./routers/StackActions').default;
+  },
+  get DrawerActions() {
+    return require('./routers/DrawerActions').default;
   },
 
   // Routers
@@ -40,27 +82,39 @@ module.exports = {
   get TabRouter() {
     return require('./routers/TabRouter').default;
   },
+  get SwitchRouter() {
+    return require('./routers/SwitchRouter').default;
+  },
 
   // Views
   get Transitioner() {
     return require('./views/Transitioner').default;
   },
-  get CardStack() {
-    return require('./views/CardStack').default;
+  get StackView() {
+    return require('./views/StackView/StackView').default;
   },
-  get Card() {
-    return require('./views/Card').default;
+  get StackViewCard() {
+    return require('./views/StackView/StackViewCard').default;
+  },
+  get SafeAreaView() {
+    return require('react-native-safe-area-view').default;
+  },
+  get SceneView() {
+    return require('./views/SceneView').default;
+  },
+  get ResourceSavingSceneView() {
+    return require('./views/ResourceSavingSceneView').default;
   },
 
   // Header
   get Header() {
-    return require('./views/Header').default;
+    return require('./views/Header/Header').default;
   },
   get HeaderTitle() {
-    return require('./views/HeaderTitle').default;
+    return require('./views/Header/HeaderTitle').default;
   },
   get HeaderBackButton() {
-    return require('./views/HeaderBackButton').default;
+    return require('./views/Header/HeaderBackButton').default;
   },
 
   // DrawerView
@@ -73,17 +127,37 @@ module.exports = {
 
   // TabView
   get TabView() {
-    return require('./views/TabView/TabView').default;
+    console.warn(
+      'TabView is deprecated. Please use the react-navigation-tabs package instead: https://github.com/react-navigation/react-navigation-tabs'
+    );
+    return require('react-navigation-deprecated-tab-navigator').TabView;
   },
   get TabBarTop() {
-    return require('./views/TabView/TabBarTop').default;
+    console.warn(
+      'TabBarTop is deprecated. Please use the react-navigation-tabs package instead: https://github.com/react-navigation/react-navigation-tabs'
+    );
+    return require('react-navigation-deprecated-tab-navigator').TabBarTop;
   },
   get TabBarBottom() {
-    return require('./views/TabView/TabBarBottom').default;
+    console.warn(
+      'TabBarBottom is deprecated. Please use the react-navigation-tabs package instead: https://github.com/react-navigation/react-navigation-tabs'
+    );
+    return require('react-navigation-deprecated-tab-navigator').TabBarBottom;
+  },
+
+  // SwitchView
+  get SwitchView() {
+    return require('./views/SwitchView/SwitchView').default;
   },
 
   // HOCs
   get withNavigation() {
     return require('./views/withNavigation').default;
+  },
+  get withNavigationFocus() {
+    return require('./views/withNavigationFocus').default;
+  },
+  get withOrientation() {
+    return require('./views/withOrientation').default;
   },
 };
