@@ -1,19 +1,19 @@
-/* @flow */
-
 import { Component } from 'react';
 import createConfigGetter from '../createConfigGetter';
 import addNavigationHelpers from '../../addNavigationHelpers';
-import type {
-  NavigationScreenOptionsGetter,
-  NavigationStackScreenOptions,
-} from '../../TypeDefinition';
+
+const dummyEventSubscriber = (name: string, handler: (*) => void) => ({
+  remove: () => {},
+});
 
 test('should get config for screen', () => {
   /* eslint-disable react/no-multi-comp */
 
   class HomeScreen extends Component {
-    static navigationOptions = ({ navigation }: *) => ({
-      title: `Welcome ${navigation.state.params ? navigation.state.params.user : 'anonymous'}`,
+    static navigationOptions = ({ navigation }) => ({
+      title: `Welcome ${
+        navigation.state.params ? navigation.state.params.user : 'anonymous'
+      }`,
       gesturesEnabled: true,
     });
 
@@ -34,7 +34,7 @@ test('should get config for screen', () => {
   }
 
   class NotificationScreen extends Component {
-    static navigationOptions = ({ navigation }: *) => ({
+    static navigationOptions = ({ navigation }) => ({
       title: '42',
       gesturesEnabled: navigation.state.params
         ? !navigation.state.params.fullscreen
@@ -46,10 +46,7 @@ test('should get config for screen', () => {
     }
   }
 
-  const getScreenOptions: NavigationScreenOptionsGetter<
-    NavigationStackScreenOptions,
-    *
-  > = createConfigGetter({
+  const getScreenOptions = createConfigGetter({
     Home: { screen: HomeScreen },
     Settings: { screen: SettingsScreen },
     Notifications: {
@@ -70,49 +67,81 @@ test('should get config for screen', () => {
 
   expect(
     getScreenOptions(
-      addNavigationHelpers({ state: routes[0], dispatch: () => false }),
+      addNavigationHelpers({
+        state: routes[0],
+        dispatch: () => false,
+        addListener: dummyEventSubscriber,
+      }),
       {}
     ).title
   ).toEqual('Welcome anonymous');
   expect(
     getScreenOptions(
-      addNavigationHelpers({ state: routes[1], dispatch: () => false }),
+      addNavigationHelpers({
+        state: routes[1],
+        dispatch: () => false,
+        addListener: dummyEventSubscriber,
+      }),
       {}
     ).title
   ).toEqual('Welcome jane');
   expect(
     getScreenOptions(
-      addNavigationHelpers({ state: routes[0], dispatch: () => false }),
+      addNavigationHelpers({
+        state: routes[0],
+        dispatch: () => false,
+        addListener: dummyEventSubscriber,
+      }),
       {}
     ).gesturesEnabled
   ).toEqual(true);
   expect(
     getScreenOptions(
-      addNavigationHelpers({ state: routes[2], dispatch: () => false }),
+      addNavigationHelpers({
+        state: routes[2],
+        dispatch: () => false,
+        addListener: dummyEventSubscriber,
+      }),
       {}
     ).title
   ).toEqual('Settings!!!');
   expect(
     getScreenOptions(
-      addNavigationHelpers({ state: routes[2], dispatch: () => false }),
+      addNavigationHelpers({
+        state: routes[2],
+        dispatch: () => false,
+        addListener: dummyEventSubscriber,
+      }),
       {}
     ).gesturesEnabled
   ).toEqual(false);
   expect(
     getScreenOptions(
-      addNavigationHelpers({ state: routes[3], dispatch: () => false }),
+      addNavigationHelpers({
+        state: routes[3],
+        dispatch: () => false,
+        addListener: dummyEventSubscriber,
+      }),
       {}
     ).title
   ).toEqual('10 new notifications');
   expect(
     getScreenOptions(
-      addNavigationHelpers({ state: routes[3], dispatch: () => false }),
+      addNavigationHelpers({
+        state: routes[3],
+        dispatch: () => false,
+        addListener: dummyEventSubscriber,
+      }),
       {}
     ).gesturesEnabled
   ).toEqual(true);
   expect(
     getScreenOptions(
-      addNavigationHelpers({ state: routes[4], dispatch: () => false }),
+      addNavigationHelpers({
+        state: routes[4],
+        dispatch: () => false,
+        addListener: dummyEventSubscriber,
+      }),
       {}
     ).gesturesEnabled
   ).toEqual(false);
@@ -135,26 +164,14 @@ test('should throw if the route does not exist', () => {
 
   expect(() =>
     getScreenOptions(
-      addNavigationHelpers({ state: routes[0], dispatch: () => false }),
+      addNavigationHelpers({
+        state: routes[0],
+        dispatch: () => false,
+        addListener: dummyEventSubscriber,
+      }),
       {}
     )
   ).toThrowError(
     "There is no route defined for key Settings.\nMust be one of: 'Home'"
   );
-});
-
-test('should throw if the screen is not defined under the route config', () => {
-  /* eslint-disable react/no-multi-comp */
-
-  const getScreenOptions = createConfigGetter({
-    Home: {},
-  });
-
-  const routes = [{ key: 'B', routeName: 'Home' }];
-
-  expect(() =>
-    getScreenOptions(
-      addNavigationHelpers({ state: routes[0], dispatch: () => false })
-    )
-  ).toThrowError('Route Home must define a screen or a getScreen.');
 });
