@@ -1,28 +1,9 @@
-/*
- * @flow
- */
-
-import invariant from 'fbjs/lib/invariant';
+import invariant from '../utils/invariant';
 
 import getScreenForRouteName from './getScreenForRouteName';
-import addNavigationHelpers from '../addNavigationHelpers';
 import validateScreenOptions from './validateScreenOptions';
 
-import type {
-  NavigationScreenProp,
-  NavigationAction,
-  NavigationRoute,
-  NavigationStateRoute,
-  NavigationRouteConfigMap,
-  NavigationScreenConfig,
-  NavigationScreenConfigProps,
-} from '../TypeDefinition';
-
-function applyConfig(
-  configurer: ?NavigationScreenConfig<*>,
-  navigationOptions: *,
-  configProps: NavigationScreenConfigProps
-): * {
+function applyConfig(configurer, navigationOptions, configProps) {
   if (typeof configurer === 'function') {
     return {
       ...navigationOptions,
@@ -41,17 +22,12 @@ function applyConfig(
   return navigationOptions;
 }
 
-export default (
-  routeConfigs: NavigationRouteConfigMap,
-  navigatorScreenConfig?: NavigationScreenConfig<*>
-) => (
-  navigation: NavigationScreenProp<NavigationRoute, NavigationAction>,
-  screenProps: *
+export default (routeConfigs, navigatorScreenConfig) => (
+  navigation,
+  screenProps
 ) => {
   const { state, dispatch } = navigation;
   const route = state;
-  // $FlowFixMe
-  const { routes, index } = (route: NavigationStateRoute);
 
   invariant(
     route.routeName && typeof route.routeName === 'string',
@@ -60,36 +36,15 @@ export default (
 
   const Component = getScreenForRouteName(routeConfigs, route.routeName);
 
-  let outputConfig = {};
-
-  if (Component.router) {
-    invariant(
-      route && routes && index != null,
-      `Expect nav state to have routes and index, ${JSON.stringify(route)}`
-    );
-    const childRoute = routes[index];
-    const childNavigation = addNavigationHelpers({
-      state: childRoute,
-      dispatch,
-    });
-    outputConfig = Component.router.getScreenOptions(
-      childNavigation,
-      screenProps
-    );
-  }
-
   const routeConfig = routeConfigs[route.routeName];
 
-  const routeScreenConfig = routeConfig.navigationOptions;
+  const routeScreenConfig =
+    routeConfig === Component ? null : routeConfig.navigationOptions;
   const componentScreenConfig = Component.navigationOptions;
 
   const configOptions = { navigation, screenProps: screenProps || {} };
 
-  outputConfig = applyConfig(
-    navigatorScreenConfig,
-    outputConfig,
-    configOptions
-  );
+  let outputConfig = applyConfig(navigatorScreenConfig, {}, configOptions);
   outputConfig = applyConfig(
     componentScreenConfig,
     outputConfig,
