@@ -1,29 +1,20 @@
-/* @flow */
-
-import invariant from 'fbjs/lib/invariant';
-
-import type {
-  NavigationComponent,
-  NavigationRouteConfigMap,
-} from '../TypeDefinition';
+import invariant from '../utils/invariant';
 
 /**
  * Simple helper that gets a single screen (React component or navigator)
  * out of the navigator config.
  */
-export default function getScreenForRouteName( // eslint-disable-line consistent-return
-  routeConfigs: NavigationRouteConfigMap,
-  routeName: string
-): NavigationComponent {
+export default function getScreenForRouteName(routeConfigs, routeName) {
   const routeConfig = routeConfigs[routeName];
 
-  invariant(
-    routeConfig,
-    `There is no route defined for key ${routeName}.\n` +
-      `Must be one of: ${Object.keys(routeConfigs)
-        .map((a: string) => `'${a}'`)
-        .join(',')}`
-  );
+  if (!routeConfig) {
+    throw new Error(
+      `There is no route defined for key ${routeName}.\n` +
+        `Must be one of: ${Object.keys(routeConfigs)
+          .map(a => `'${a}'`)
+          .join(',')}`
+    );
+  }
 
   if (routeConfig.screen) {
     return routeConfig.screen;
@@ -41,5 +32,5 @@ export default function getScreenForRouteName( // eslint-disable-line consistent
     return screen;
   }
 
-  invariant(false, `Route ${routeName} must define a screen or a getScreen.`);
+  return routeConfig;
 }
