@@ -1,61 +1,56 @@
-/** @flow */
-
-import invariant from 'fbjs/lib/invariant';
-
-import type { NavigationRouteConfigMap } from '../TypeDefinition';
+import invariant from '../utils/invariant';
 
 /**
  * Make sure the config passed e.g. to StackRouter, TabRouter has
  * the correct format, and throw a clear error if it doesn't.
  */
-function validateRouteConfigMap(routeConfigs: NavigationRouteConfigMap) {
+function validateRouteConfigMap(routeConfigs) {
   const routeNames = Object.keys(routeConfigs);
   invariant(
     routeNames.length > 0,
     'Please specify at least one route when configuring a navigator.'
   );
 
-  routeNames.forEach((routeName: string) => {
+  routeNames.forEach(routeName => {
     const routeConfig = routeConfigs[routeName];
+    const screenComponent = getScreenComponent(routeConfig);
 
-    invariant(
-      routeConfig.screen || routeConfig.getScreen,
-      `Route '${routeName}' should declare a screen. ` +
-        'For example:\n\n' +
-        "import MyScreen from './MyScreen';\n" +
-        '...\n' +
-        `${routeName}: {\n` +
-        '  screen: MyScreen,\n' +
-        '}'
-    );
-
-    if (routeConfig.screen && routeConfig.getScreen) {
-      invariant(
-        false,
-        `Route '${routeName}' should declare a screen or ` +
-          'a getScreen, not both.'
-      );
-    }
-
-    if (routeConfig.screen) {
-      invariant(
-        typeof routeConfig.screen === 'function',
+    if (
+      !screenComponent ||
+      (typeof screenComponent !== 'function' &&
+        typeof screenComponent !== 'string' &&
+        !routeConfig.getScreen)
+    ) {
+      throw new Error(
         `The component for route '${routeName}' must be a ` +
-          'a React component. For example:\n\n' +
+          'React component. For example:\n\n' +
           "import MyScreen from './MyScreen';\n" +
           '...\n' +
-          `${routeName}: {\n` +
-          '  screen: MyScreen,\n' +
+          `${routeName}: MyScreen,\n` +
           '}\n\n' +
           'You can also use a navigator:\n\n' +
           "import MyNavigator from './MyNavigator';\n" +
           '...\n' +
-          `${routeName}: {\n` +
-          '  screen: MyNavigator,\n' +
+          `${routeName}: MyNavigator,\n` +
           '}'
       );
     }
+
+    if (routeConfig.screen && routeConfig.getScreen) {
+      throw new Error(
+        `Route '${routeName}' should declare a screen or ` +
+          'a getScreen, not both.'
+      );
+    }
   });
+}
+
+function getScreenComponent(routeConfig) {
+  if (!routeConfig) {
+    return null;
+  }
+
+  return routeConfig.screen ? routeConfig.screen : routeConfig;
 }
 
 export default validateRouteConfigMap;
