@@ -1,37 +1,14 @@
-/* @flow */
-
-import React, { PureComponent } from 'react';
+import React from 'react';
 import propTypes from 'prop-types';
+import { NavigationProvider } from './NavigationContext';
 
-import type {
-  NavigationScreenProp,
-  NavigationRoute,
-  NavigationAction,
-  NavigationNavigatorProps,
-} from '../TypeDefinition';
-
-type Props = {
-  screenProps?: {},
-  navigation: NavigationScreenProp<NavigationRoute, NavigationAction>,
-  component: ReactClass<NavigationNavigatorProps<NavigationRoute>>,
-};
-
-export default class SceneView extends PureComponent<void, Props, void> {
-  static childContextTypes = {
-    navigation: propTypes.object.isRequired,
-  };
-
-  props: Props;
-
-  getChildContext() {
-    return {
-      navigation: this.props.navigation,
-    };
-  }
-
+export default class SceneView extends React.PureComponent {
   render() {
-    const { screenProps, navigation, component: Component } = this.props;
-
-    return <Component screenProps={screenProps} navigation={navigation} />;
+    const { screenProps, component: Component, navigation } = this.props;
+    return (
+      <NavigationProvider value={navigation}>
+        <Component screenProps={screenProps} navigation={navigation} />
+      </NavigationProvider>
+    );
   }
 }
