@@ -1,5 +1,3 @@
-/* @flow */
-
 /**
  * TouchableItem renders a touchable that looks native on both iOS and Android.
  *
@@ -9,34 +7,19 @@
  * On iOS you can pass the props of TouchableOpacity, on Android pass the props
  * of TouchableNativeFeedback.
  */
-import React, { Component, Children } from 'react';
+import React from 'react';
 import {
   Platform,
   TouchableNativeFeedback,
   TouchableOpacity,
   View,
 } from 'react-native';
-import type { Style } from '../TypeDefinition';
 
 const ANDROID_VERSION_LOLLIPOP = 21;
 
-type Props = {
-  onPress: Function,
-  delayPressIn?: number,
-  borderless?: boolean,
-  pressColor?: ?string,
-  activeOpacity?: number,
-  children?: React.Element<*>,
-  style?: Style,
-};
-
-type DefaultProps = {
-  pressColor: ?string,
-};
-
-export default class TouchableItem
-  extends Component<DefaultProps, Props, void> {
+export default class TouchableItem extends React.Component {
   static defaultProps = {
+    borderless: false,
     pressColor: 'rgba(0, 0, 0, .32)',
   };
 
@@ -53,8 +36,7 @@ export default class TouchableItem
       Platform.OS === 'android' &&
       Platform.Version >= ANDROID_VERSION_LOLLIPOP
     ) {
-      const { style, ...rest } = this.props; // eslint-disable-line no-unused-vars
-
+      const { style, ...rest } = this.props;
       return (
         <TouchableNativeFeedback
           {...rest}
@@ -64,17 +46,13 @@ export default class TouchableItem
             this.props.borderless
           )}
         >
-          <View style={this.props.style}>
-            {Children.only(this.props.children)}
-          </View>
+          <View style={style}>{React.Children.only(this.props.children)}</View>
         </TouchableNativeFeedback>
       );
     }
 
     return (
-      <TouchableOpacity {...this.props}>
-        {this.props.children}
-      </TouchableOpacity>
+      <TouchableOpacity {...this.props}>{this.props.children}</TouchableOpacity>
     );
   }
 }
