@@ -1,19 +1,18 @@
-/* @flow */
-
 import ScenesReducer from '../ScenesReducer';
 
 /**
  * Simulate scenes transtion with changes of navigation states.
  */
-function testTransition(states: *) {
-  const routes = states.map((keys: *) => ({
+function testTransition(states) {
+  const routes = states.map(keys => ({
     index: 0,
-    routes: keys.map((key: *) => ({ key, routeName: '' })),
+    routes: keys.map(key => ({ key, routeName: '' })),
+    isTransitioning: false,
   }));
 
   let scenes = [];
   let prevState = null;
-  routes.forEach((nextState: *) => {
+  routes.forEach(nextState => {
     scenes = ScenesReducer(scenes, nextState, prevState);
     prevState = nextState;
   });
@@ -91,17 +90,18 @@ describe('ScenesReducer', () => {
     const state1 = {
       index: 0,
       routes: [{ key: '1', routeName: '' }, { key: '2', routeName: '' }],
+      isTransitioning: false,
     };
 
     const state2 = {
       index: 1,
       routes: [{ key: '1', routeName: '' }, { key: '2', routeName: '' }],
+      isTransitioning: false,
     };
 
     const scenes1 = ScenesReducer([], state1, null);
     const scenes2 = ScenesReducer(scenes1, state2, state1);
-    /* $FlowFixMe: We want tests to fail on undefined */
-    const route = scenes2.find((scene: *) => scene.isActive).route;
+    const route = scenes2.find(scene => scene.isActive).route;
     expect(route).toEqual({ key: '2', routeName: '' });
   });
 
@@ -109,11 +109,13 @@ describe('ScenesReducer', () => {
     const state1 = {
       index: 0,
       routes: [{ key: '1', routeName: '' }, { key: '2', routeName: '' }],
+      isTransitioning: false,
     };
 
     const state2 = {
       index: 0,
       routes: [{ key: '1', routeName: '' }, { key: '2', routeName: '' }],
+      isTransitioning: false,
     };
 
     const scenes1 = ScenesReducer([], state1, null);
@@ -125,11 +127,13 @@ describe('ScenesReducer', () => {
     const state1 = {
       index: 0,
       routes: [{ key: '1', routeName: '' }, { key: '2', routeName: '' }],
+      isTransitioning: false,
     };
 
     const state2 = {
       index: 0,
       routes: [{ key: '2', routeName: '' }, { key: '1', routeName: '' }],
+      isTransitioning: false,
     };
 
     const scenes1 = ScenesReducer([], state1, null);
@@ -144,6 +148,7 @@ describe('ScenesReducer', () => {
         { key: '1', x: 1, routeName: '' },
         { key: '2', x: 2, routeName: '' },
       ],
+      isTransitioning: false,
     };
 
     const state2 = {
@@ -152,6 +157,7 @@ describe('ScenesReducer', () => {
         { key: '1', x: 3, routeName: '' },
         { key: '2', x: 4, routeName: '' },
       ],
+      isTransitioning: false,
     };
 
     const scenes1 = ScenesReducer([], state1, null);
@@ -166,6 +172,7 @@ describe('ScenesReducer', () => {
         { key: '1', x: 1, routeName: '' },
         { key: '2', x: 2, routeName: '' },
       ],
+      isTransitioning: false,
     };
 
     const state2 = {
@@ -174,6 +181,7 @@ describe('ScenesReducer', () => {
         { key: '1', x: 1, routeName: '' },
         { key: '2', x: 2, routeName: '' },
       ],
+      isTransitioning: false,
     };
 
     const scenes1 = ScenesReducer([], state1, null);
